Remove allocate_flattened_array helper function
When using the conda `gfortran_linux-64` package (gcc 7.2.0), the pointer
used in the BMI `get_value` function is already associated, and therefore
is allocatable. I don't think this was the case in the older conda `gcc`
package (gcc 4.8.5). This change affects the `get_value`, `get_value_ref`,
and `get_value_at_indices` functions.