Change prefix of deferred procedures  from 'bmi_' to 'bmif_'
This will allow the use of names such as `bmi_initialize` at a later
point, if desired.

The `bmi_` prefix is only used for procedures in an interface that
defers implentation *with a different name* (i.e., the actual BMI
function name, like `initialize`) until the abstract type in which
they're defined is extended, so there's no harm in changing the prefix
to match the library name.