Output an array, not a pointer, from get_value
A BMI should not allocate memory. Instead, when using, e.g.,
`get_value`, the caller should create a variable, allocating memory to
hold array output. The variable is passed as a parameter into
`get_value`, where values are copied into it. Fortran arrays make this
easy. For example, in `get_grid_shape`, these statements

    shape(1) = self%model%n_y
    shape(2) = self%model%n_x

and this statement

    shape = [self%model%n_y, self%model%n_x]

are equivalent, elementwise copies into the variable `shape`.

Use this concept in the `get_value` methods, outputting an array, not
a pointer, with the values of the specified variable.