Use BMI version in module name
This will help prevent collisions when the BMI is installed. The module
name is Fortran identifier, so it can't include hyphens or underscores.
Compiling the module will result in a file named "bmif_1_2.mod". The
module will be imported with

    use bmif_1_2

which is ugly, but clear. Plus, the "bmi" type included within the module
is unaffected.